backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
sqlalchemy>=2.0.22
python-dotenv>=1.0.0
python-multipart>=0.0.6
pypdf2>=3.0.1
google-generativeai>=0.3.1
psycopg2-binary>=2.9.9  # For PostgreSQL (Neon)
